TEXT EXTRACTION FROM AN IMAGE
In this project, we will be working on extracting text from images. After extracting the text we will apply some basic functions of OpenCV on that text to enhance it and to get more accurate results. This project will be very useful as it will save time and effort of typing from an image.
Different organization can use this to extract useful information from the image and store it. Individuals can use it for saving their time and effort for typing.